Huge refactor of V2. Not complete
