--- V4/Constants/ErrorMessages.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IpswichJAFFARunningClubAPI\V4\Constants;
-
-abstract class ErrorMessages
-{
-    const GENERIC_ERROR_MESSAGE = 'Unknown error in reading results from the database';
-}
