--- V4/IRoute.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace IpswichJAFFARunningClubAPI\V4;
-
-interface IRoute
-{
-  public function registerRoutes();
-}
