--- V4/Results/ResultsDataAccess.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace IpswichJAFFARunningClubAPI\V4\Results;
-
-require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V4/DataAccess.php';
-
-use IpswichJAFFARunningClubAPI\V4\DataAccess as DataAccess;
-
-class ResultsDataAccess extends DataAccess
-{
-}
