+++ v2/Constants/Rules.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace IpswichJAFFARunningClubAPI\V2\Constants;
+
+abstract class Rules
+{
+    const GENERIC_ERROR_MESSAGE = 'Unknown error in reading results from the database';
+    const START_OF_2015_AGE_GRADING = '2017-01-01';
+}
