+++ v2/Rankings/RankingsController.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace IpswichJAFFARunningClubAPI\V2\Rankings;
+
+require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/BaseController.php';
+require_once IPSWICH_JAFFA_API_PLUGIN_PATH . 'V2/IRoute.php';
+require_once 'RankingsDataAccess.php';
+
+use IpswichJAFFARunningClubAPI\V2\BaseController as BaseController;
+use IpswichJAFFARunningClubAPI\V2\IRoute as IRoute;
+
+class RankingsController extends BaseController implements IRoute
+{
+	public function __construct(string $route, $db)
+	{
+		parent::__construct($route, new RankingsDataAccess($db));
+	}
+
+	public function registerRoutes()
+	{
+	
+	}
+}
